Make it possible to insert initial records
The foreign key constraint on the `next_id` column meant that we can't insert
the first record, because any ID we put on next_id won't be a valid foreign key
(since there's nothing in the table).

This migration removes that foreign key constraint so that we can insert the
first few records and kickstart the ring.