Always break the ring at the first element
Previously we were doing some, uh, interesting stuff around newest/oldest
Redirections, which assumed that the newest redirection links to the oldest one.
That's not necessarily true.

Now we:

* Grab the first Redirection (`@from_redirection`)
* Grab the Redirection that that Redirection points to (`@to_redirection`)
* Break the link between them
* Insert a new redirection between them

This ensures that we're breaking a link that definitely exists, instead of breaking a link
that we assume (based on creation time) exists.

This commit also refactors the tests to use the existing gabe/edward Redirections
that are created via DB seeds.