Create atom feed of webring members
This is a relatively simple feed that shows "gabebw joined Hotline
Webring" with a link to the site that joined.

Atom feeds require an author. I added a new key to our en.yml file for this
because someone's gotta author these amazing feed entries.

Also, according to the [atom
standard](http://tools.ietf.org/html/rfc4287#page-20) entries in an
atom feed should never, ever change. They should also not be the same
as the URI (like the website they're pointing to) to limit confusion as to what
they are used for. They are a unique string used for identification.

`hotlinewebring.club` is hardcoded into for tag URI formula so that even if the
app's domain changes, the unique, never changing ids will stay the
same. Hopefully.

Here is a blog post for reference:
https://web.archive.org/web/20081225232147/http://diveintomark.org/archi
ves/2004/05/28/howto-atom-id