Use created_at in atom feed entries
When a new Redirection is added to the webring, the first Redirection in the ring points to it. This changes that first Redirection's `updated_at` value.

In the atom feed, we were using `updated_at` to say when that entry was updated. While that's semantic, it would cause the first Redirection to continue to pop up in atom feeds because it looked like there was new information. No longer!

We now use `created_at` in that field, because for the purposes of the feed, we only want to see each Redirection when it's added, not "updated".