Add a test for visiting a High Voltage page
On production, starting the server immediately fails with this error:

    undefined method `hide_action' for HighVoltage::PagesController:Class

This is because `hide_action` was removed from controllers in Rails 5.

This error does _not_ occur in development or test because those environments
have `config.eager_load` set to `false`, so the `HighVoltage::PagesController`
code is never loaded, and thus doesn't fail.

The test added in this commit exercises the `HighVoltage::PagesController` code,
and fails without upgrading high_voltage.

To fix this issue, upgrade to high_voltage 3.0.0, which supports Rails 5:
https://github.com/thoughtbot/high_voltage/blob/master/NEWS.md