Add a method to remove a site from the webring
I know, I know: who would want such a thing?!

The `unlink` method runs in a transaction so that removing a Redirection and
re-linking its neighbors is an atomic operation. In other words, we can't remove
a Redirection, run into an error, and then leave a hole in the ring.