Add link_audit rake task
This taks should help us to maintenance on the webring's members. It
checks each redirection for proper next and previous links. It downloads
the page and scans the raw html for the proper urls. It doesn't check
for actual links because some sites (looking at you @jnutting) don't use
normal <a href="..."> links.

This also found other issues while I was testing it like sites that have
migrated to https because the html only returned the redirection (and
therefore didn't include the links).