Store the whole original referring URL
We currently store only the hostname of the HTTP referrer (e.g.
`gabebw.com/posts/2018/whatever` is saved as just `gabebw.com`).

This is a problem for pages on free sites, where the homepage for a
given user is not the root hostname but a subdirectory (e.g.
`freepages.rootsweb.com/~gabebw`).

This is not a fix for issue #58, but it does allow us to manually fix
that issue if it recurs.