Make the link audit task more resilient
* Just like before, if there's a `SocketError`, indicate that the
  website is completely offline
* If there is any other kind of error (`StandardError`), print the error
  message

The certificate for https://irongeek.net expired and the `rake
link_audit` task crashed when it got to it. Now it prints the error and
keeps going.

To accommodate showing more information, the `MissingLinkFinder` now
sends more information back to the link audit task, including a `status`
plus information relating to that status.