Update sprockets-rails gem for Ruby 2.7
We need this PR to prevent warnings from Ruby 2.7:
https://github.com/rails/sprockets-rails/pull/454

However, the last sprockets-rails release was in 2017, so we grab it
from GitHub. It's pinned to the earliest ref that includes that PR so
that we don't always grab the latest master (which is a shifting target
and could introduce instability).