Log HTTP headers on redirection failure
We've had a few times where we wanted to see some evidence of redirection creation or failure, but heroku doesn't log headers (like the HTTP referrer).

This adds logging for when we fail to find or create the redirection so we have a little more idea about what might have gone wrong in the redirection creation process.

We log HTTP headers only. `request.env` or `request.headers` has a lot of other info in it besides HTTP headers. If we just grab headers that start with HTTP we get only HTTP headers. We also don't want to to see the cookie header because it's got a ton of obfuscated info that we don't really care about.