Correctly match blocked referrer paths
Our previous code used simple string comparison, which meant that if
`evil.com/subpath` was blocked, we would also block everything under
`evil.com/subpathABC` (since it starts with `evil.com/subpathABC`).

The fix is to use code that understands that the URLs are actually URLs,
not simple strings, so that we can correctly compare the hosts and
paths.

This uses `Pathname#ascend` (docs:
https://ruby-doc.org/stdlib-2.5.3/libdoc/pathname/rdoc/Pathname.html#method-i-ascend)
which yields an iterator over the parent directories of a given
Pathname.

For example, given `/foo/bar/baz`, `ascend` yields `/foo/bar/baz`, then
`/foo/bar`, then `/foo`, then `/`. The yielded values are Pathnames, not
strings.