Redirect to an existing blocked Redirection
If the Redirection is already in the webring, then redirect to it
instead of to `Redirection.first`.

This fixes a bug with the Redirection for HLWR itself: we blocked the
`"hotlinewebring.club"` URL so that nobody can add it as their own site,
but because it's blocked, it doesn't redirect to its own next/previous
links.

This change also means that when we block but do not unlink a site, then
clicking on its links will redirect to its actual next/previous links
instead of `Redirection.first`, which is fine. Blocked sites cannot be
added to the ring, but if they're already in, then they work normally.