Block subdomains of an existing redirection
If we already have `foo.com`, we now block `bar.foo.com` from joining.
Unfortunately we have to do `Redirection.all.find` to do this check.
It's not great, but it's probably not that bad in practice since we
don't have thousands of Redirections.

If we could add arbitrary extensions to Postgres, I'd do it inside
Postgres with https://github.com/petere/pguri and compare `uri_host`
plus some string manipulation, but we can't do that.