Only validate new users when they try to join
Most of our `before_action`s are about validating *new* users. But most
of our traffic is *existing* users who already have a `Redirection` and
we just need to follow it.

For those existing users, validating the slug, referrer, etc is a waste
of resources. Now we skip doing that work.