Pass Heroku app name in plain text
For some reason, the action can't find the secrets. To fix it, just pass
them in plain text. This is the error we were getting:

    Error: Command failed: heroku git:remote --app
    Error: Flag --app expects a value

This indicates that the `heroku_app_name` value is blank for some
reason.

You can see it on https://github.com/hotline-webring/hotline-webring/actions/runs/4440720997