Allow us to close registration (#403)
We're overwhelmed with new registrations. This allow us to set the
environment variable `CLOSED_FOR_REGISTRATION=1` which will allow all
existing registrations to work, but redirect all new registrations to a
"closed" page.

It also updates the homepage to show that we're closed so people can
easily see before signing up.

Note that this is different from the
`DISALLOW_CREATING_NEW_REDIRECTIONS` config, which is more like a
shadow ban when we're getting a bunch of spam. It does not show a
message. Surprisingly different implementations!