Remove video
This pun used to call us on our cell phone
Late-night in two thousand sixteen

but then we made better moderation rules and grew up a little and
decide that the song is a little sexist which is clearly against our
rules.

That can only mean one thing

Again, remove the video from the website.