Add last seen to redirections
We have a lot of sites now and we don't check every one all the time, so
they can get go offline and we won't notice. Sometimes, kind webring
members will point it out to us, but it would be nicer if we could be
proactive about it.

We've tried to automate this in the past with tools, but when things
render with javascript, or in iframes, it gets tougher to do.

Instead, what if we updated the redirection every time someone clicked
the link? That might give us enough data to quickly see which sites are
inactive/offline.