Add Redirection.order_nulls_last
Now that we are sorting by `last_seen_at`, and that column can be
blank, we will be sorting by `null`s. If we sort so that recently seen
redirections are on top, all the redirections we _never_ "seen" will
show up first.

Instead, we can ues postgres' [`NULLS LAST`][nulls-last] functionality
which will always put them at the end, in `null` jail.

[nulls-last]: https://www.postgresql.org/docs/17/queries-order.html